requirements.txt
tensorflow>=2.0.0
keras
Flask
Pillow
Werkzeug
numpy
matplotlib
scipy
gunicorn
